Logarithmic Transformation of Fluorescence Data

Applies a custom logarithmic transformation to fluorescence intensity data, transforming values greater than 1 using a base-10 logarithm and setting values less than or equal to 1 to zero.

Usage

LogSingleData(x)

Arguments

x

A numeric vector of fluorescence intensity values to be transformed.

Value

A numeric vector containing the transformed fluorescence intensity values.

Details

This function is used internally to transform fluorescence intensity data. It ensures that non-positive values (which cannot be log-transformed) are set to zero, while positive values greater than 1 are log-transformed using base 10. This helps in handling data with a wide range of values and avoids issues with undefined logarithms.

Examples

if (FALSE) { # \dontrun{
transformed_values <- LogSingleData(fluorescence_values)
} # }