Generate basic QC plots for Tocky data

This function visualizes either Timer fluorescence (Blue vs Red) or Timer dynamics by the Tocky method (Angle vs Intensity) based on the specified mode.

Usage

plot_tocky(
  x,
  file = "PlotTocky",
  pseudocolour = TRUE,
  pdf = FALSE,
  output = "QC",
  n = 4,
  plot_mode = "Timer fluorescence",
  lower_quantile_cutoff = 0.01,
  select = FALSE,
  group = TRUE,
  group_order = NULL,
  interactive = TRUE,
  save = FALSE,
  samplefile = NULL,
  verbose = TRUE
)

Arguments

x

A TockyPrepData object returned by `timer_transform`, which sample grouping has been defined by `sampledef`.

file

File name.

pseudocolour

A logical argument for whether to use pseudocolour in plots.

pdf

A logical argument; if FALSE, a jpeg file is generated instead.

output

The output directory name for output files.

n

A number; n x n plots will be generated in the output Tocky plot file, max is 4 for 16 plots.

plot_mode

Either "Timer fluorescence" for Blue vs Red plots, "Normalized Timer fluorescence" for normalized plots, or "Timer Angle and Intensity" for Angle vs Intensity plots.

lower_quantile_cutoff

Lower quantile cutoff for setting the plot ranges in fluorescence mode.

select

Logical indicating whether to manually select samples for plotting.

group

Logical indicating whether to group plots based on the `group` field in `sampledef`.

group_order

Optional character vector for specifying the order of the panels when using the group option.

interactive

Logical indicating whether to prompt the user to select plot_mode. Defaults to `TRUE`.

save

A logical argument; if FALSE, plots are shown in an X window.

samplefile

Character vector specifying the sample files. Defaults to `NULL`.

verbose

Logical indicating whether to print progress messages. Default is `TRUE`.

Examples

if (FALSE) { # \dontrun{
  plot_tocky(x, plot_mode = "Timer fluorescence")
  plot_tocky(x, plot_mode = "Timer Angle and Intensity")
} # }