Plot Timer Data for Shiny Visualization

This function provides visualizations for timer data within a Shiny application, showing different modes such as "Timer fluorescence", "Normalized Timer fluorescence", or "Timer Angle and Intensity". It is designed to be used internally by Shiny apps to dynamically display changes based on user input parameters.

Usage

plot_tocky_shiny(
  x,
  plot_mode = "Timer fluorescence",
  lower_quantile_cutoff = 0.01
)

Arguments

x

A `TockyPrepData` object from `timer_transform` containing processed timer data.

plot_mode

A character string indicating the type of plot to display. Options are "Timer fluorescence", "Normalized Timer fluorescence", or "Timer Angle and Intensity". Defaults to "Timer fluorescence".

lower_quantile_cutoff

A numeric value indicating the lower quantile cutoff for setting the plot ranges in fluorescence mode. This helps in adjusting the scale of the plot to focus on relevant data points. Defaults to 0.01.

Value

This function does not return a value but renders a plot directly to the Shiny application. It adjusts plot elements such as labels, axis limits, and coloring based on the input parameters.

Details

The function checks if the input data inherits from `TockyPrepData`. It then uses base R plotting functions to create plots according to the specified `plot_mode`. The function adjusts the x and y limits based on the provided `lower_quantile_cutoff` and uses density colors for better visualization of data point distributions.

Examples

if (FALSE) { # \dontrun{
  plot_tocky_shiny(x, plot_mode = "Timer fluorescence")
} # }