Prepare Data for Timer Transformation Using Flow Cytometric Data

This function prepares the dataset for timer transformation analysis by identifying common variables across sample files, configuring necessary control files, and setting up variables for downstream analysis. The function supports both interactive and non-interactive file selection modes.

Usage

prep_tocky(path = ".", interactive = TRUE, negfile = NULL, samplefile = NULL)

Arguments

path

Character string specifying the directory where the data files are located. Defaults to the current directory `'.'`.

interactive

Logical indicating whether to prompt the user to select sample files. Defaults to `TRUE`.

negfile

Character string specifying the negative control file. If `NULL`, the user will be prompted to select a file. Defaults to `NULL`.

samplefile

Character vector specifying the sample files. If `NULL` and `samplefilechoice` is `TRUE`, the user will be prompted to select files. Defaults to `NULL`.

Value

A list containing paths to the control file, selected sample files, and the standardized variables used in the analysis.

Examples

if (FALSE) { # \dontrun{
  # Interactive file selection
  prep_data <- prep_tocky(path='data', output='output')

  # Specifying files directly for non-interactive usage
  prep_data <- prep_tocky(
    path='data',
    output='output',
    negfile='neg_control.csv',
    samplefile=c('sample1.csv', 'sample2.csv')
  )
} # }