Calculate Quadrant Gating Logic

Determines which quadrant each point in a dataset falls into based on specified x and y thresholds.

Usage

quadrant_gate(df, x_point, y_point, var_x_name, var_y_name)

Arguments

df

A data frame containing the fluorescence data.

x_point

The numeric threshold for the x-axis (e.g., Red fluorescence).

y_point

The numeric threshold for the y-axis (e.g., Blue fluorescence).

var_x_name

The name of the x-axis variable in `df`.

var_y_name

The name of the y-axis variable in `df`.

Value

A list containing logical vectors for each quadrant:

Q1: Points below `x_point` and above or equal to `y_point`.

Q2: Points above or equal to `x_point` and above or equal to `y_point`.

Q3: Points above or equal to `x_point` and below `y_point`.

Q4: Points below `x_point` and below `y_point`.

Details

This function is intended for internal use to assist in categorizing cells into quadrants based on fluorescence gating thresholds. It helps in further analysis such as determining the percentage of cells within each quadrant.

Examples

if (FALSE) { # \dontrun{
  df <- data.frame(Red = rnorm(100, 50, 10), Blue = rnorm(100, 50, 10))
  quadrant_data <- quadrant_gate(df, x_point = 50, y_point = 50,
  var_x_name = "Red", var_y_name = "Blue")
} # }