Update sample definitions and group assignments

This function takes the output from `timer_transform`, specifically the `sample_definition` data frame, exports it to a CSV file for the user to edit group assignments, and then reads the updated file back into R.

Usage

sample_definition(
  timer_transform_output,
  sample_definition = NULL,
  output_dir = NULL,
  filename = "sample_definition.csv",
  sep = ",",
  verbose = TRUE,
  interactive = FALSE
)

Arguments

timer_transform_output

A TockyPrepData object returned by `timer_transform`, containing `sample_definition`.

sample_definition

(Optional) to use a data frame object as an input, specify the input object by this parameter. Defaul is `NULL`.

output_dir

Character string specifying the directory to save the `sample_definition.csv` file. If `NULL`, the file is saved in the current working directory. Default is `NULL`.

filename

Character string specifying the name of the sample definition file. Default is `"sample_definition.csv"`.

sep

Character string indicating the field separator used in the CSV file. Default is `","`.

verbose

Logical indicating whether to display messages. Default is `TRUE`.

interactive

Logical indicating whether to use an interactive session to export a file for sample grouping and enable user to edit it and import. Defaults to `TRUE`.

Value

An updated TockyPrepData with user-assigned groupings.

Examples

if (FALSE) { # \dontrun{
  # Assuming `result` is the output from `timer_transform`
  sample_def <- sample_definition(result, output_dir = "output_directory")
  # The function will pause, allowing you to edit the 'group' column in the CSV file.
  # After editing and saving the file, press Enter in R to continue.
  # The updated sample definitions will be returned as a data frame.
} # }